requirements.txt
discord
jishaku
discord-pretty-help
psutil
pyyaml
py-cpuinfo
aiosqlite
